# [ENH] faster groupby top k