# PR Description

Please describe the changes proposed in the pull request:

- Use nlargest/nsmallest instead ... avoids sorting the entire dataframe
- change column names to align more with pandas description and other existing functions (open to suggestions, or possible revert to existing names)
- deprecate sort_kwargs, since there is no use of `sort_values`
-  add `ascending` parameter to match existing parameter in the deprecated `sort_kwargs`

<!-- Doing so provides maintainers with context on what the PR is, and can help us more effectively review your PR. -->

<!-- Please also identify below which issue that has been raised that you are going to close. -->

**This PR resolves #1093 .**


# performance comparison ... view with a pinch of salt, as always

```py
# adapted from  https://stackoverflow.com/a/36702926/7175713
np.random.seed(3)
df = pd.DataFrame({
   'dep': np.random.choice(list('ABC'), 2_000_000),
  'manager_id': np.random.randint(0, 200_000, 2_000_000),
   'salary': np.random.randint(5000, 5006, 2_000_000)
 })

df.head()
  dep  manager_id  salary
0   C       12586    5002
1   A      153719    5004
2   B      178720    5004
3   A      162938    5001
4   A      122360    5000

# --------------------> PR

df.groupby_topk('dep', 'salary', 5)
   dep  manager_id  salary
0    C       44168    5000
1    C      135457    5000
2    C       42892    5000
3    C      130785    5000
4    C      155488    5000
5    A      122360    5000
6    A      136722    5000
7    A       14219    5000
8    A      162417    5000
9    A      115906    5000
10   B       66927    5000
11   B       79314    5000
12   B      150702    5000
13   B      186580    5000
14   B       34819    5000

df.groupby_topk('dep', 'salary', 5, ignore_index = False) 
    dep  manager_id  salary
11    C       44168    5000
13    C      135457    5000
42    C       42892    5000
73    C      130785    5000
118   C      155488    5000
4     A      122360    5000
14    A      136722    5000
30    A       14219    5000
46    A      162417    5000
49    A      115906    5000
29    B       66927    5000
33    B       79314    5000
38    B      150702    5000
51    B      186580    5000
76    B       34819    5000

%timeit df.groupby_topk('dep', 'salary', 5)
162 ms ± 12.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

%timeit df.groupby_topk('dep', 'salary', 5, ignore_index = False)
154 ms ± 1.5 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)



# ---------------------------> dev

df.groupby_topk('dep', 'salary', 5)
            dep  manager_id  salary
dep                                
A   1999996   A      131800    5000
    548719    A       57184    5000
    548720    A       59279    5000
    1585713   A       87340    5000
    548727    A      196661    5000
B   716069    B       89977    5000
    306266    B       34652    5000
    743267    B      125335    5000
    1877756   B      166187    5000
    1703182   B      101396    5000
C   1999997   C      180787    5000
    292147    C      105282    5000
    1609388   C       52935    5000
    1609401   C       55700    5000
    292136    C      111160    5000

df.groupby_topk('dep', 'salary', 5).droplevel(0) 
        dep  manager_id  salary
1999996   A      131800    5000
548719    A       57184    5000
548720    A       59279    5000
1585713   A       87340    5000
548727    A      196661    5000
716069    B       89977    5000
306266    B       34652    5000
743267    B      125335    5000
1877756   B      166187    5000
1703182   B      101396    5000
1999997   C      180787    5000
292147    C      105282    5000
1609388   C       52935    5000
1609401   C       55700    5000
292136    C      111160    5000

%timeit df.groupby_topk('dep', 'salary', 5)
297 ms ± 4.18 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
 ```


<!-- As you go down the PR template, please feel free to delete sections that are irrelevant. -->

# PR Checklist

<!-- This checklist exists for newcomers who are not yet familiar with our requirements. If you are experienced with
the project, please feel free to delete this section. -->

Please ensure that you have done the following:

1. [ ] PR in from a fork off your branch. Do not PR from `<your_username>`:`dev`, but rather from `<your_username>`:`<feature-branch_name>`.
<!-- Doing this helps us keep the commit history much cleaner than it would otherwise be. -->
2. [x] If you're not on the contributors list, add yourself to `AUTHORS.md`.
<!-- We'd like to acknowledge your contributions! -->
3. [x] Add a line to `CHANGELOG.md` under the latest version header (i.e. the one that is "on deck") describing the contribution.
    - Do use some discretion here; if there are multiple PRs that are related, keep them in a single line.

# Automatic checks

There will be automatic checks run on the PR. These include:

- Building a preview of the docs on Netlify
- Automatically linting the code
- Making sure the code is documented
- Making sure that all tests are passed
- Making sure that code coverage doesn't go down.

# Relevant Reviewers

<!-- Finally, please tag relevant maintainers to review. -->

Please tag maintainers to review.

- @ericmjl