@samukweku 2022-05-16
made the update with the old code; not sure if I'm missing something else. by the way @thatlittleboy how do you format the docs with black? or do you copy the code snippet, run it in a .py and copy back?