@thatlittleboy 2022-05-16
> made the update with the old code; not sure if I'm missing something else. by the way @thatlittleboy how do you format the docs with black? or do you copy the code snippet, run it in a .py and copy back?

I don't typically run `black` if the snippets are small enough to "fix by hand" (in general, I know what the "black" style should look like, so I just write in that manner when coding up the MWEs).
There might be tools like `blacken-docs` that we can consider integrating, though that one doesn't seem to support the way of declaring code snippets that we are using.